restarted

+++ backend/src/users/users.controller.ts
@@ -9,21 +9,22 @@ import {
   ParseIntPipe,
   Post,
   Query,
+  Req,
 } from '@nestjs/common';
 import { Prisma, User } from '@prisma/client';
-import { UsersService } from './users.service';
-import { ZodValidationPipe } from 'src/zod.validation.pipe';
+import { AccesRoles } from 'src/acces-roles/acces-roles.decorator';
+import { Role } from 'src/acces-roles/role.enum';
 import { CreateUserDto, CreateUserSchema } from 'src/dto/create-user.dto';
 import { PaginationDto, PaginationSchema } from 'src/dto/pagination.dto';
+import { ZodValidationPipe } from 'src/zod.validation.pipe';
+import { UsersService } from './users.service';
+import { ModifyUserDto, ModifyUserSchema } from 'src/dto/modify-user.dto';
 
 @Controller('users')
 export class UsersController {
   constructor(private usersService: UsersService) {}
-  // @Get()
-  // getAllUsers(): Promise<User[]> {
-  //   return this.usersService.getUsers({});
-  // }
 
+  @AccesRoles(Role.Admin, Role.User)
   @Get()
   async getUsers(
     @Query(new ZodValidationPipe(PaginationSchema))
@@ -38,7 +39,7 @@ export class UsersController {
     return this.usersService.getUser({ id: id });
   }
 
-  @Post()
+  @Post('create')
   async createUser(
     @Body(new ZodValidationPipe(CreateUserSchema)) createUserDto: CreateUserDto,
   ): Promise<User> {
@@ -64,14 +65,53 @@ export class UsersController {
     }
   }
 
+  @AccesRoles(Role.Admin, Role.User)
   @Delete(':id')
-  async deleteUser(@Param('id', ParseIntPipe) id: number): Promise<User> {
+  async deleteUser(
+    @Param('id', ParseIntPipe) id: number,
+    @Req() context,
+  ): Promise<User> {
+    if (context.user.role === Role.User && context.user.id !== id) {
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+    }
     try {
-      return await this.usersService.deleteUser({ id: id });
-    } catch {
+      return this.usersService.deleteUser({ id: id });
+    } catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        if (error.code === 'P2025') {
+          // No record found
+          throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+      }
       throw new HttpException(
-        'User with this id does not exist',
-        HttpStatus.NOT_FOUND,
+        'Internal server error',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+  }
+
+  @Post('modify')
+  async changeUserData(
+    @Body(new ZodValidationPipe(ModifyUserSchema)) modifyUserDto: ModifyUserDto,
+  ): Promise<User> {
+    console.log(modifyUserDto);
+    const id = modifyUserDto.id;
+    delete modifyUserDto.id;
+    try {
+      return this.usersService.updateUser({
+        where: { id: id },
+        data: modifyUserDto as Prisma.UserUpdateInput,
+      });
+    } catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError) {
+        if (error.code === 'P2025') {
+          // No record found
+          throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+      }
+      throw new HttpException(
+        'Internal server error',
+        HttpStatus.INTERNAL_SERVER_ERROR,
       );
     }
   }
